refactored registerUser
