+++ src/main/java/net/crusadergames/bugwars/model/auth/Role.java
@@ -1,5 +1,6 @@
 package net.crusadergames.bugwars.model.auth;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,13 +15,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Role {
+    @JsonIgnore
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
-
-  @Enumerated(EnumType.STRING)
-  @Column(length = 20)
-  private ERole name;
-
+    @Enumerated(EnumType.STRING)
+    @Column(length = 20)
+    private ERole name;
 }
