+++ src/main/java/net/crusadergames/bugwars/repository/auth/UserRepository.java
@@ -1,4 +1,4 @@
-package net.crusadergames.bugwars.repository;
+package net.crusadergames.bugwars.repository.auth;
 
 import net.crusadergames.bugwars.model.auth.User;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,9 +8,9 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-  Optional<User> findByUsername(String username);
+    Optional<User> findByUsername(String username);
 
-  Boolean existsByUsername(String username);
+    Boolean existsByUsername(String username);
 
-  Boolean existsByEmail(String email);
+    Boolean existsByEmail(String email);
 }
