+++ src/main/java/net/crusadergames/bugwars/service/AuthService.java
@@ -1,0 +1,77 @@
+package net.crusadergames.bugwars.service;
+
+import net.crusadergames.bugwars.model.auth.ERole;
+import net.crusadergames.bugwars.model.auth.Role;
+import net.crusadergames.bugwars.model.auth.User;
+import net.crusadergames.bugwars.payload.request.LoginRequest;
+import net.crusadergames.bugwars.payload.request.SignupRequest;
+import net.crusadergames.bugwars.payload.response.JwtResponse;
+import net.crusadergames.bugwars.repository.auth.RoleRepository;
+import net.crusadergames.bugwars.repository.auth.UserRepository;
+import net.crusadergames.bugwars.security.jwt.JwtUtils;
+import net.crusadergames.bugwars.security.service.UserDetailsImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+@Service
+public class AuthService {
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    RoleRepository roleRepository;
+
+    @Autowired
+    PasswordEncoder passwordEncoder;
+
+    @Autowired
+    AuthenticationManager authenticationManager;
+
+    @Autowired
+    JwtUtils jwtUtils;
+
+    public ResponseEntity<?> authenticateUser(LoginRequest loginRequest) {
+        Authentication authentication = authenticationManager
+                .authenticate(new UsernamePasswordAuthenticationToken(loginRequest.getUsername(), loginRequest.getPassword()));
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        String jwt = jwtUtils.generateJwtToken(authentication);
+
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+        List<String> roles = userDetails.getAuthorities().stream().map(GrantedAuthority::getAuthority)
+                .collect(Collectors.toList());
+
+        return ResponseEntity
+                .ok(new JwtResponse(jwt, userDetails.getId(), userDetails.getUsername(), roles));
+    }
+
+    public User registerUser(SignupRequest signUpRequest) {
+        if (userRepository.existsByUsername(signUpRequest.getUsername())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Username is already taken");
+        }
+
+        if (userRepository.existsByEmail(signUpRequest.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Email is already taken");
+        }
+
+        User user = new User(signUpRequest.getUsername(), signUpRequest.getEmail(),
+                passwordEncoder.encode(signUpRequest.getPassword()));
+        Role userRole = roleRepository.findByName(ERole.ROLE_USER)
+                .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+        user.setRoles(Set.of(userRole));
+        return userRepository.save(user);
+    }
+}
