+++ src/main/resources/application.properties
@@ -1,17 +1,20 @@
 spring.config.import=optional:file:env.properties
 
 spring.datasource.url=jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
-spring.datasource.username= ${DB_USER}
-spring.datasource.password= ${DB_PASSWORD}
+spring.datasource.username=${DB_USER}
+spring.datasource.password=${DB_PASSWORD}
 
 spring.jpa.defer-datasource-initialization=true
 spring.sql.init.mode=${INIT_MODE}
-spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation= true
-spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.open-in-view=true
 
 # Hibernate ddl auto (create, create-drop, validate, update)
-spring.jpa.hibernate.ddl-auto= ${DB_DDL}
+spring.jpa.hibernate.ddl-auto=${DB_DDL}
 
-jwtSecret= ${JWT_SECRET}
-jwtExpirationMs= ${JWT_EXPIRATION_MS}
+server.error.include-message=always
+server.error.include-stacktrace=never
+
+jwtSecret=${JWT_SECRET}
+jwtExpirationMs=${JWT_EXPIRATION_MS}
